+++ Boids/Models/Position.cs
@@ -5,9 +5,9 @@ namespace Boids.Models
     public class Position
     {
         [JsonProperty("x")]
-        public float X { get; set; }
+        public double X { get; set; }
 
         [JsonProperty("y")]
-        public float Y { get; set; }
+        public double Y { get; set; }
     }
 }
