+++ Boids/Models/Velocity.cs
@@ -5,6 +5,6 @@ namespace Boids.Models
     public class Velocity : Position
     {
         [JsonProperty("deg")]
-        public float Deg { get; set; }
+        public double Deg { get; set; }
     }
 }
